Revert "fixed entity, .properiest converted to .yaml"
This reverts commit 42defab8fc3d9776cb5e87e30a408e1a62e8016e.

+++ src/main/java/ru/emiljan/servicedevdevices/models/Image.java
@@ -3,6 +3,7 @@ package ru.emiljan.servicedevdevices.models;
 import lombok.*;
 
 import javax.persistence.*;
+import java.util.List;
 
 /**
  * @author EM1LJAN
@@ -17,7 +18,7 @@ import javax.persistence.*;
 public class Image {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(name = "icon64")
     private String icon64;
